Added MarketingButtons component test cases
