+++ src/components/marketing/Marketing.test.js
@@ -4,7 +4,6 @@ import { shallow, mount } from 'enzyme';
 import { createShallow } from '@material-ui/core/test-utils';
 import Marketing from './Marketing';
 import AppBar from '@material-ui/core/AppBar';
-import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Drawer from '@material-ui/core/Drawer';
 import Hidden from '@material-ui/core/Hidden';
@@ -117,37 +116,5 @@ describe('<Marketing /> component testing', () => {
        });
      });
 
-     describe('Marketing container testing',() => {
-           it('should render Marketing container correctly',() => {
-                 const mainContainer = findByTestAttr(component, 'marketing-container');
-                 expect(mainContainer.length).toBe(1);
-           });
-
-           it('should render second container div correctly',() => {
-            const secondContainer = findByTestAttr(component, 'marketing-second-container');
-            expect(secondContainer.length).toBe(1);
-      });
-        
-         it('should render container title correctly', () => {
-               const h3 = findByTestAttr(component, 'container-title');
-               expect(h3.length).toBe(1);
-               expect(h3.text()).toEqual('Connect with Mentors in your area!');
-         });
-         it('should render buttons container correctly', () => {
-               const buttonsContainer = findByTestAttr(component, 'buttons-container');
-               expect(buttonsContainer.length).toBe(1);
-         });
-
-         it('should render Sign Up button with no errors', () => {
-               const signUpButton = findByTestAttr(component,'buttons-container').find(Button).at(0);
-               expect(signUpButton.exists()).toBe(true);
-               expect(signUpButton.text()).toEqual('Sign Up');
-         });
-
-         it('should render Login button with no errors', () => {
-                  const logInButton = findByTestAttr(component,'buttons-container').find(Button).at(1);
-                  expect(logInButton.exists()).toBe(true);
-                  expect(logInButton.text()).toEqual('Login');
-            });
-      });  
+     
 });
